rename memtable_flush to mem_table_flush
