+++ src/lib.rs
@@ -15,7 +15,7 @@ mod flatbuffer_types;
 mod flush;
 mod iter;
 mod mem_table;
-mod memtable_flush;
+mod mem_table_flush;
 mod sst;
 mod sst_iter;
 mod tablestore;
